tests/wpt/web-platform-tests/content-security-policy/script-src/script-src-1_4_2.html

<!DOCTYPE HTML>
<html>
<head>
    <title>Function() called as a constructor should throw without 'unsafe-eval' script-src directive.</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
</head>
<body>
    <h1>Function() called as a constructor should throw without 'unsafe-eval' script-src directive.</h1>
    <div id='log'></div>

   	<script>

        test(function() {
            assert_throws(
                new EvalError(),
                function() {
                    var funq = new Function('');
                    funq();
        })}, "Unsafe eval ran in Function() constructor.");

    </script>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=script-src%20%27self%27+%27unsafe-inline%27'></script>

</body>
</html>
